app/requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
argon2-cffi-bindings==21.2.0; python_version >= '3.6'
argon2-cffi==21.3.0
asgiref==3.4.1; python_version >= '3.6'
bcrypt==3.2.0
beautifulsoup4==4.10.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9; python_version >= '3'
cryptography==36.0.1
defusedxml==0.7.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
django-allauth==0.47.0
django-boost==1.7.2
django-bootstrap4==21.1
django-cleanup==5.2.0
django-croppie==1.3
django-debug-toolbar==3.2.4
django-environ==0.8.1
django-extra-views==0.14.0
django[argon2]==3
factory-boy==3.2.1
faker==10.0.0; python_version >= '3.6'
gunicorn==20.1.0
idna==3.3; python_version >= '3'
numpy==1.21.4
oauthlib==3.1.1; python_version >= '3.6'
opencv-python==4.5.4.60
pillow==8.4.0
psycopg2-binary==2.8.6
pycparser==2.21; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pyjwt[crypto]==2.3.0; python_version >= '3.6'
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python3-openid==3.2.0
pytz==2021.3
requests-oauthlib==1.3.0
requests==2.26.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
setuptools==59.7.0
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sorl-thumbnail==12.7.0
soupsieve==2.3.1; python_version >= '3.6'
sqlparse==0.4.2; python_version >= '3.5'
text-unidecode==1.3
ua-parser==0.10.0
urllib3==1.26.7; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
user-agents==2.2.0
wheel==0.37.0
